requirements/base.txt
# Core
Django==2.2.*
django-configurations==2.1

# For the persistence stores
psycopg2-binary==2.8.3
dj-database-url==0.5.0


# Rest apis
djangorestframework==3.9.*
django-cors-headers==3.0.2
